requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
jinja2==3.1.4
pydantic==2.7.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
# Templates handling
python-docx==1.1.2
docxtpl==0.19.0
openpyxl==3.1.5
pypdf==4.3.1
reportlab==4.2.2
jsonschema==4.23.0
pypdfium2==4.30.0
# For calling soffice to convert to PDF
psutil==6.0.0